Add most of the logic for the addon.
- Refactor the logic to work with multiple type of procs (critical,
deflect, no shield -- last one is not implemented yet)
- Add a second proc window to be able to show 2 procs at once. Should be
enough for now until we add the NoShield logic. We will need a 3rd proc
window to be able to have all 3 warrior procs
- Add the logic to display the spell icon in the proc window when triggered
- Add a FinishAddon() method to clean the EventHandler if we are started
on a class we don't manage (Esper for example)
- Add the logic to hide the proc window if the corresponding spell is
currently in cooldown. We want to improve that in the future to show
the cooldown in the proc window though. or potentially have a setting to
choose which solution.
- Utility method to get a spell from an id or to get the cooldown of a
spell

TESTED
- Everything needs to be tested and debugged
- Cooldown and GetSpellId should work as we tested them in TestCooldown
and GetSpellId projects before.
- Need to make sure the UI is valid in Houston. It was modified directly
in the XML...

IMPROVEMENTS TO MAKE:

TODO:
* Add cooldown logic => v2: show with cooldown overlay
* Add cooldown logic => v3: make it possible to switch between both
implementation (hidden or overlay) with an option

* Or make the proc window movable (see CandyBars)
  - Should be enabled by an option

* Show a counter on the proc window showing how long the proc is still
active

* Options form: unlock movable frame
* Options form: activate/deactivate the addon for specific spells.
Should show only the spells of your class
* Options form: cooldown logic (hidden / overlay)
* Options form: enable/disable the proc timer